*** empty log message ***

+++ load_ogg.c
@@ -43,7 +43,7 @@ static size_t sdl_read_func(void *ptr, size_t size, size_t nmemb, void *datasour
 
 static int sdl_seek_func(void *datasource, ogg_int64_t offset, int whence)
 {
-    return SDL_RWseek((SDL_RWops*)datasource, offset, whence);
+    return SDL_RWseek((SDL_RWops*)datasource, (int)offset, whence);
 }
 
 static int sdl_close_func_freesrc(void *datasource)
@@ -105,7 +105,7 @@ SDL_AudioSpec *Mix_LoadOGG_RW (SDL_RWops *src, int freesrc,
     spec->freq = info->rate;
     spec->samples = 4096; /* buffer size */
     
-    samples = ov_pcm_total(&vf, -1);
+    samples = (long)ov_pcm_total(&vf, -1);
 
     *audio_len = spec->size = samples * spec->channels * 2;
     *audio_buf = malloc(*audio_len);
